Moved some dungeon mob classes around.
