+++ src/main/java/com/gildedgames/aether/client/renderer/entity/layers/SentryLayer.java
@@ -2,7 +2,7 @@ package com.gildedgames.aether.client.renderer.entity.layers;
 
 import com.gildedgames.aether.Aether;
 
-import com.gildedgames.aether.common.entity.monster.SentryEntity;
+import com.gildedgames.aether.common.entity.monster.dungeon.SentryEntity;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.IEntityRenderer;
 import net.minecraft.client.renderer.entity.layers.AbstractEyesLayer;
