+++ src/main/java/com/gildedgames/aether/client/renderer/entity/model/MimicModel.java
@@ -1,6 +1,6 @@
 package com.gildedgames.aether.client.renderer.entity.model;
 
-import com.gildedgames.aether.common.entity.monster.MimicEntity;
+import com.gildedgames.aether.common.entity.monster.dungeon.MimicEntity;
 import com.mojang.blaze3d.matrix.MatrixStack;
 import com.mojang.blaze3d.vertex.IVertexBuilder;
